+++ js/application.js
@@ -1,3 +1,4 @@
+
 import introScreen from './screen/intro-screen';
 import greetingScreen from './screen/greeting-screen';
 import rulesScreen from './screen/rules-screen';
