+++ js/main.js
@@ -1,3 +1,4 @@
+
 import Application from './application';
 
 Application.showIntro();
