+++ js/screen/greeting-screen.js
@@ -1,3 +1,4 @@
+
 import AbstractView from '../view';
 import Application from '../application';
 
@@ -15,7 +16,6 @@ class GreetingView extends AbstractView {
         <div class="greeting__continue"><span><img src="img/arrow_right.svg" width="64" height="64" alt="Next"></span></div>
       </div>`;
   }
-
 
   bindHandlers() {
     this.element.querySelector('.greeting__continue').onclick = () => Application.showRules();
