+++ js/view.js
@@ -1,3 +1,4 @@
+
 export default class AbstractView {
 
   get element() {
